Replace extra whitespace moved to shared code
