+++ src/applications/appeals/shared/constants.js
@@ -81,7 +81,7 @@ export const MAX_LENGTH = {
  **** REGULAR EXPRESSIONS ****
  */
 export const REGEXP = {
-  WHITESPACE: /\s/g,
+  WHITESPACE: /\s+/g,
   APOSTROPHE: /\u2019/g,
   PERCENT: /(\s|\b)percent(\s|\b)/gi,
 };
