+++ src/applications/appeals/shared/tests/utils/replace.unit.spec.js
@@ -4,7 +4,7 @@ import {
   replaceDescriptionContent,
   replaceSubmittedData,
   fixDateFormat,
-} from '../../../10182/utils/replace';
+} from '../../utils/replace';
 
 describe('replaceDescriptionContent', () => {
   it('should return an empty string', () => {
@@ -13,8 +13,8 @@ describe('replaceDescriptionContent', () => {
     expect(replaceDescriptionContent('')).to.eq('');
   });
   it('should not alter an these strings', () => {
-    expect(replaceDescriptionContent('   ')).to.eq('   ');
     expect(replaceDescriptionContent('abc 123')).to.eq('abc 123');
+    expect(replaceDescriptionContent('a b c 123')).to.eq('a b c 123');
   });
   it('should replace percent with a % symbol', () => {
     expect(replaceDescriptionContent('10 percent')).to.eq('10%');
@@ -29,6 +29,10 @@ describe('replaceDescriptionContent', () => {
     expect(replaceDescriptionContent('20 percentile')).to.eq('20 percentile');
     expect(replaceDescriptionContent('percentpercent')).to.eq('percentpercent');
   });
+  it('should strip out multiple whitespace & trim', () => {
+    expect(replaceDescriptionContent('   ')).to.eq('');
+    expect(replaceDescriptionContent(' a  b  c  \t  d')).to.eq('a b c d');
+  });
 });
 
 describe('replaceSubmittedData', () => {
@@ -37,15 +41,19 @@ describe('replaceSubmittedData', () => {
     expect(replaceSubmittedData(null)).to.eq('');
     expect(replaceSubmittedData('')).to.eq('');
   });
-  it('should not alter an these strings', () => {
-    expect(replaceSubmittedData('   ')).to.eq('   ');
+  it('should not alter these strings', () => {
     expect(replaceSubmittedData('abc 123')).to.eq('abc 123');
+    expect(replaceSubmittedData('a b c 123')).to.eq('a b c 123');
   });
   it('should replace typographical apostrophes with a single quote', () => {
     expect(replaceSubmittedData('don’t won’t can’t you’ll')).to.eq(
       "don't won't can't you'll",
     );
     expect(replaceSubmittedData('’100’ times')).to.eq("'100' times");
+  });
+  it('should strip out multiple whitespace & trim', () => {
+    expect(replaceSubmittedData('   ')).to.eq('');
+    expect(replaceSubmittedData(' a  b  c  \t  d')).to.eq('a b c d');
   });
 });
 
@@ -70,5 +78,7 @@ describe('fixDateFormat', () => {
     expect(fixDateFormat('2020- 1 - 2 ')).to.eq('2020-01-02');
     expect(fixDateFormat('2023 - 10 - 1')).to.eq('2023-10-01');
     expect(fixDateFormat('2000-6 - 30')).to.eq('2000-06-30');
+    expect(fixDateFormat('2000  -  6 - 30')).to.eq('2000-06-30');
+    expect(fixDateFormat('2000 \t - \t 6 \t - \t 30')).to.eq('2000-06-30');
   });
 });
